requirements.txt
Flask==3.0.3
flask_socketio==5.3.6
openai==1.24.0
openpyxl==3.1.2
pandas==2.2.2
pinecone==0.1.0
python-dotenv==1.0.1
ratelimit==2.2.1
Requests==2.31.0
